Clean up some code and change class purposes ipfsservice <-> ipfsfilemanager
