+++ opendb-api/src/main/resources/application.yml
@@ -10,17 +10,13 @@ server:
     max-http-post-size: 4096
     max-threads: 5
 
-ipfs:
-  run: ${IPFS_RUN:true}
-  storing:
-    time: ${IPFS_STORING_TIME:86400}
-  host: ${IPFS_HOST:localhost} #dev host: 51.158.69.207
-  port: ${IPFS_PORT:5001}
-  timeout: ${IPFS_TIMEOUT:10000}
-  directory: ${IPFS_DIRECTORY:/opendb/storage/}
-  cluster:
-    host: ${IPFS_CLUSTER_HOST:localhost}
-    port: ${IPFS_CLUSTER_PORT:9094}
+storage:
+  ipfs.node:
+    host: ${IPFS_HOST:}
+    port: ${IPFS_PORT:5001}
+    readTimeoutMs: ${IPFS_READ_TIMEOUT_MS:10000}
+  local-storage: ${EXT_RES_DIRECTORY:}
+  timeToStoreUnusedSec: ${EXT_RES_STORE_UNUSED:86400}
 
 opendb:
   mgmt:
@@ -33,6 +29,13 @@ opendb:
   db:
     compactCoefficient: ${OPENDB_DB_CF:1.0}
     dbSuperblockSize: ${OPENDB_DB_BLOCK_SZ:32}
+  storage:
+    ipfs.node:
+      host: ${IPFS_HOST:}
+      port: ${IPFS_PORT:5001}
+      readTimeoutMs: ${IPFS_READ_TIMEOUT_MS:10000}
+    local-storage: ${EXT_RES_DIRECTORY:}
+    timeToStoreUnusedSec: ${EXT_RES_STORE_UNUSED:86400}
   db-schema:
     objtables:
       obj_logins: { "types": ["sys.login", "sys.signup"], "keysize":2 }
