+++ opendb-core/src/main/java/org/openplacereviews/opendb/api/IPFSController.java
@@ -3,9 +3,10 @@ package org.openplacereviews.opendb.api;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.openplacereviews.opendb.service.ipfs.IPFSService;
-import org.openplacereviews.opendb.service.ipfs.dto.ImageDTO;
-import org.openplacereviews.opendb.service.ipfs.dto.IpfsStatusDTO;
+import org.openplacereviews.opendb.dto.IpfsStatusDTO;
+import org.openplacereviews.opendb.dto.ResourceDTO;
+import org.openplacereviews.opendb.service.IPFSFileManager;
+import org.openplacereviews.opendb.service.IPFSService;
 import org.openplacereviews.opendb.util.JsonFormatter;
 import org.openplacereviews.opendb.util.exception.ConnectionException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,39 +32,49 @@ public class IPFSController {
 	protected static final Log LOGGER = LogFactory.getLog(IPFSController.class);
 
 	@Autowired
-	private IPFSService ipfsService;
+	private IPFSFileManager externalResourcesManager;
 
 	@Autowired
 	private JsonFormatter formatter;
+	
+	private void checkIPFSRunning() {
+		if(!externalResourcesManager.isRunning()) {
+			throw new ConnectionException("IPFS service is not running.");
+		}
+	}
 
 	@PostMapping(value = "/image", consumes = MULTIPART_FORM_DATA_VALUE, produces = "text/html;charset=UTF-8")
 	@ResponseBody
 	public ResponseEntity<String> uploadImage(
 			@RequestPart(name = "file") @Valid @NotNull @NotEmpty MultipartFile file) throws IOException {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		ImageDTO imageDTO = ImageDTO.of(file);
-		imageDTO = ipfsService.addFile(imageDTO);
+		checkIPFSRunning();
+		// TODO add to file system and to ipfs
+		ResourceDTO imageDTO = ResourceDTO.of(file);
+		imageDTO = externalResourcesManager.addFile(imageDTO);
 
 		return ResponseEntity.ok(formatter.imageObjectToJson(imageDTO));
 	}
 
-	@GetMapping
 	@ResponseBody
+	@GetMapping("status")
 	public ResponseEntity<String> loadIpfsStatus() throws IOException, UnirestException {
-		IpfsStatusDTO ipfsStatusDTO = IPFSService.status ? ipfsService.getIpfsNodeInfo() : getIpfsStatus();
-
+		// TODO return json with all information
+		// TODO add parameter full and display all counts of pending, in blockchain objects etc
+		IpfsStatusDTO ipfsStatusDTO ;
+		if(!externalResourcesManager.isRunning()) {
+			ipfsStatusDTO = new IpfsStatusDTO().setStatus("NOT CONNECTED");
+		} else {
+			ipfsStatusDTO = externalResourcesManager.getIpfsNodeInfo(); 
+		}
 		return ResponseEntity.ok(formatter.fullObjectToJson(ipfsStatusDTO));
 	}
+	
 
 	@GetMapping(value = "/image")
 	@ResponseBody
 	public ResponseEntity<String> getFile(@RequestParam("cid") String cid) throws IOException {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		try (ByteArrayOutputStream outputStream = (ByteArrayOutputStream) ipfsService.read(cid)) {
+		checkIPFSRunning();
+		try (ByteArrayOutputStream outputStream = (ByteArrayOutputStream) externalResourcesManager.read(cid)) {
 			return ResponseEntity.ok(Base64.getEncoder().encodeToString(outputStream.toByteArray()));
 		}
 	}
@@ -71,23 +82,20 @@ public class IPFSController {
 	@PutMapping(value = "/image")
 	@ResponseBody
 	public ResponseEntity<String> pinImage(@RequestParam("cid") String cid) {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		if (!ipfsService.pin(cid)) {
+		checkIPFSRunning();
+		if (!externalResourcesManager.pin(cid)) {
 			return ResponseEntity.ok("{\"status\":\"FAILED\"}");
 		}
 
 		return ResponseEntity.ok("{\"status\":\"OK\"}");
 	}
+	
 
 	@DeleteMapping(value = "/image")
 	@ResponseBody
 	public ResponseEntity<String> unpinImage(@RequestParam("cid") String cid) {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		if(!ipfsService.unpin(cid)) {
+		checkIPFSRunning();
+		if(!externalResourcesManager.unpin(cid)) {
 			return ResponseEntity.ok("{\"status\":\"FAILED\"}");
 		}
 
@@ -97,51 +105,37 @@ public class IPFSController {
 	@GetMapping(value = "/image/tracked")
 	@ResponseBody
 	public ResponseEntity<String> getTrackedFiles() {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		List<String> tracked = ipfsService.getTracked();
+		checkIPFSRunning();
+		List<String> tracked = externalResourcesManager.getPinnedResources();
 		return ResponseEntity.ok(formatter.fullObjectToJson(tracked));
 	}
 
 	@GetMapping(value = "/status")
 	@ResponseBody
 	public ResponseEntity<String> getStatusCheckingMissingImagesInIPFS() {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		return ResponseEntity.ok(formatter.fullObjectToJson(ipfsService.checkingMissingImagesInIPFS()));
+		checkIPFSRunning();
+		return ResponseEntity.ok(formatter.fullObjectToJson(externalResourcesManager.checkingMissingImagesInIPFS()));
 	}
 
-	@PutMapping(value = "/upload")
+	@PostMapping(value = "/mgmt/ipfs-maintenance")
 	@ResponseBody
-	public ResponseEntity<String> loadMissingImagesToIpfs() {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		return ResponseEntity.ok(formatter.fullObjectToJson(ipfsService.uploadMissingImagesToIPFS()));
+	public ResponseEntity<String> uploadMissingImagesToIPFS() {
+		checkIPFSRunning();
+		return ResponseEntity.ok(formatter.fullObjectToJson(externalResourcesManager.uploadMissingResourcesToIPFS()));
 	}
 
-	@GetMapping(value = "/blc-status")
+	@GetMapping(value = "/mgmt/ipfs-status")
 	@ResponseBody
 	public ResponseEntity<String> getStatusCheckingMissingImagesInBlockchain() {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		return ResponseEntity.ok(formatter.fullObjectToJson(ipfsService.statusImagesInDB()));
+		checkIPFSRunning();
+		return ResponseEntity.ok(formatter.fullObjectToJson(externalResourcesManager.statusImagesInDB()));
 	}
 
-	@DeleteMapping(value = "/blc-images")
+	@PostMapping(value = "/mgmt/clean-deprecated-ipfs")
 	@ResponseBody
 	public ResponseEntity<String> removeUnactivatedAndTimeoutImages() throws IOException {
-		if (!IPFSService.status)
-			throw new ConnectionException("IPFS service was not runned!");
-
-		return ResponseEntity.ok(formatter.fullObjectToJson(ipfsService.removeUnusedImageObjectsFromSystemAndUnpinningThem()));
-	}
-
-	private IpfsStatusDTO getIpfsStatus() {
-		return new IpfsStatusDTO().setStatus("NOT CONNECTED");
+		checkIPFSRunning();
+		return ResponseEntity.ok(formatter.fullObjectToJson(externalResourcesManager.removeUnusedImageObjectsFromSystemAndUnpinningThem()));
 	}
 
 }
