+++ opendb-core/src/main/java/org/openplacereviews/opendb/ops/OpOperation.java
@@ -1,10 +1,20 @@
 package org.openplacereviews.opendb.ops;
 
-import com.google.gson.*;
-import org.openplacereviews.opendb.service.ipfs.dto.ImageDTO;
-
 import java.lang.reflect.Type;
-import java.util.*;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
 
 public class OpOperation extends OpObject {
 	
@@ -23,7 +33,6 @@ public class OpOperation extends OpObject {
 	
 	private List<OpObject> newObjects = new LinkedList<OpObject>();
 	protected String type;
-	protected List<ImageDTO> images;
 	
 	public OpOperation() {
 	}
@@ -129,14 +138,6 @@ public class OpOperation extends OpObject {
 		return getStringValue(F_COMMENT);
 	}
 
-	public List<ImageDTO> getImages() {
-		return images;
-	}
-
-	public void setImages(List<ImageDTO> images) {
-		this.images = images;
-	}
-
 	@Override
 	public int hashCode() {
 		final int prime = 31;
@@ -191,10 +192,6 @@ public class OpOperation extends OpObject {
 		public OpOperation deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
 				throws JsonParseException {
 			Map treeMap = new Gson().fromJson(json, TreeMap.class);
-
-			List<ImageDTO> imageDTOList = new ArrayList<>();
-			getImageObject(treeMap, imageDTOList);
-
 			JsonObject jsonObj = json.getAsJsonObject();
 			OpOperation op = new OpOperation();
 			JsonElement tp = jsonObj.remove(F_TYPE);
@@ -213,7 +210,6 @@ public class OpOperation extends OpObject {
 			}
 
 			jsonObj.remove(F_EVAL);
-			op.images = imageDTOList;
 			op.fields = context.deserialize(jsonObj, TreeMap.class);
 			return op;
 		}
@@ -232,41 +228,7 @@ public class OpOperation extends OpObject {
 			return context.serialize(tm);
 		}
 
-		private void getImageObject(Map map, List<ImageDTO> array) {
-			if (map.containsKey(F_TYPE) && map.get(F_TYPE).equals("#image")) {
-				array.add(ImageDTO.of(map.get("hash").toString(), map.get("extension").toString(), map.get("cid").toString()));
-			} else {
-				map.keySet().forEach(key -> {
-					if (map.get(key) instanceof Map) {
-						getImageObject( (Map) map.get(key), array);
-					}
-					if (map.get(key) instanceof List) {
-						if (!(((List) map.get(key)).isEmpty()) && !(((List) map.get(key)).get(0).getClass().equals(String.class))) {
-							getImageObject( (List<Map>)map.get(key), array);
-						}
-					}
-				});
-			}
-		}
-
-		private void getImageObject(List<Map> list, List<ImageDTO> array) {
-			for (Map map : list) {
-				map.keySet().forEach(key -> {
-					if (key.equals(F_TYPE) && map.get(key).equals("#image")) {
-						array.add(ImageDTO.of(map.get("hash").toString(), map.get("extension").toString(), map.get("cid").toString()));
-					} else {
-						if (map.get(key) instanceof Map) {
-							getImageObject( (Map) map.get(key), array);
-						}
-						if (map.get(key) instanceof List) {
-							if (map.get(key).getClass().equals(Map.class)) {
-								getImageObject((List<Map>) map.get(key), array);
-							}
-						}
-					}
-				});
-			}
-		}
+		
 
 	}
 }
