+++ opendb-core/src/main/java/org/openplacereviews/opendb/service/BlocksManager.java
@@ -8,7 +8,6 @@ import org.openplacereviews.opendb.SecUtils;
 import org.openplacereviews.opendb.api.MgmtController;
 import org.openplacereviews.opendb.ops.*;
 import org.openplacereviews.opendb.ops.OpBlockchainRules.ErrorType;
-import org.openplacereviews.opendb.service.ipfs.IPFSService;
 import org.openplacereviews.opendb.util.JsonFormatter;
 import org.openplacereviews.opendb.util.OUtils;
 import org.openplacereviews.opendb.util.exception.FailedVerificationException;
@@ -187,24 +186,10 @@ public class BlocksManager {
 				return null;
 			}
 		}
-
-		if (IPFSService.status) {
-			candidates.forEach(operation -> {
-				if (!operation.getImages().isEmpty()) {
-					operation.getImages().forEach(imageDTO -> {
-						dataManager.updateImageActiveStatus(imageDTO, true);
-						ipfsService.getReplicaSet().forEach(cluster -> {
-							try {
-								cluster.pin(imageDTO.getCid());
-							} catch (UnirestException e) {
-								e.printStackTrace();
-							}
-						});
-					});
-				}
-			});
-		}
 		timer.measure(tmAddOps, ValidationTimer.BLC_ADD_OPERATIONS);
+		
+		ipfsService.processOperations(candidates);
+		timer.measure(tmAddOps, ValidationTimer.BLC_PROCESS_RESOURCES);
 
 		int tmNewBlock = timer.startExtra();
 		OpBlock opBlock = blc.createBlock(serverUser, serverKeyPair);
